Revert "Monitoring Angular directives to use React components (#19183)"
This reverts commit 2216b226d6c1bf0b10317bef0d31a597ef12a5e2.
