+++ x-pack/plugins/monitoring/public/components/elasticsearch/indices/indices.js
@@ -11,7 +11,7 @@ import { LARGE_FLOAT, LARGE_BYTES, LARGE_ABBREVIATED } from '../../../../common/
 import { formatMetric } from '../../../lib/format_number';
 import { ElasticsearchStatusIcon } from '../status_icon';
 import { ClusterStatus } from '../cluster_status';
-import { MonitoringTable } from '../../table';
+import { MonitoringTable } from '../../';
 import { EuiLink } from '@elastic/eui';
 import { KuiTableRowCell, KuiTableRow } from '@kbn/ui-framework/components';
 import { SystemIndicesCheckbox } from './system_indices_checkbox';
