+++ x-pack/plugins/monitoring/public/components/elasticsearch/overview/overview.js
@@ -7,7 +7,7 @@
 import React, { Fragment } from 'react';
 import { ClusterStatus } from '../cluster_status';
 import { ShardActivity } from '../shard_activity';
-import { MonitoringTimeseriesContainer } from '../../chart';
+import { MonitoringTimeseriesContainer } from 'plugins/monitoring/components';
 
 export function ElasticsearchOverview({
   clusterStatus,
