+++ x-pack/plugins/monitoring/public/components/index.js
@@ -4,13 +4,13 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-
-/*
- * This file should only export page-level components for view controllers to
- * mount React to the DOM
- */
+export { MonitoringTimeseriesContainer } from './chart';
+export { MonitoringTable } from './table';
+export { Tooltip } from './tooltip';
 export { NoData } from './no_data';
 export { License } from './license';
+export { StatusIcon } from './status_icon';
+export { SummaryStatus } from './summary_status'; // TODO this line can be removed, component is only used by other components
 export { PageLoading } from './page_loading';
 export {
   ElasticsearchOverview,
