+++ x-pack/plugins/monitoring/public/directives/all.js
@@ -8,6 +8,7 @@ import './main';
 import './chart';
 import './sparkline';
 import './alerts';
+import './status_icon';
 import './cluster/overview';
 import './cluster/listing';
 import './elasticsearch/cluster_status';
